refactor 5. 인기검색어와 실시간검색어 컴포넌트 분리
- 제곧내.
